+++ src/lib/socket.ts
@@ -24,6 +24,9 @@ export enum SocketEvent {
   BILL_STATUS_CHANGED = 'bill:status-changed',
   VENDOR_CREATED = 'vendor:created',
   VENDOR_UPDATED = 'vendor:updated',
+  VENDOR_ACCOUNT_CREATED = 'vendor:account:created',
+  VENDOR_ACCOUNT_UPDATED = 'vendor:account:updated',
+  VENDOR_ACCOUNT_DELETED = 'vendor:account:deleted',
   COMMENT_ADDED = 'comment:added',
   ATTACHMENT_ADDED = 'attachment:added',
 }